Fixed ActIndex Crowbar error with the lockpick Fixed other people not being able to use vehicles when use protection is on.
Added a convar to teamban everyone automatically from every team, then admins can unban people from teams(Everyone is a citizen until unbanned by admin)
This is ridiculous and should not be turned on. This doesn't get saved either and is such a worthless addition I only added it because some guy was constantly fucking asking me to do it.