DAMN this was hard, I made it so you can see the owner of a prop, and also that you can see whether you can touch it with the weapon you're holding.
If you're holding the physgun then it will show a tick/cross at the middle left of the screen if you can pick it up with the physgun
same for toolgun/gravgun.

if you're not holding any of those three guns it will check if you can damage it. This needs extended bug testing since I had to change about everything to get this to work. Please report bugs in the thread or DarkRP googlecode page.