FAdmin is now really compatible with ULX!
FAdmin setroot yourname
And you will be root_user in both ULX AND FAdmin. ULX has root_user when FAdmin is installed. And the root_user in ULX has the same access as superadmin. But in FPP the root_user has privilege giving access while superadmin does not.


Shit is cache :D