Drakehawke and I have come to an agreement.
The current form of the barricade is too intrusive and leading. I'm reverting it until I've changed it to the following:
When someone bans me they simply get a confirmation box with a text explaining who I am. They won't get an "FPtje ban fail" message and they won't be called a prick when they click "yes". 

This way it's not really intrusive, it doesn't make it harder to ban me and it will generally have a better tone.

I hope you guys can agree.